BUG (string): Series.str.slice with negative step (#59724)
